// client/models/student/basicInfo.model.ts
export interface BasicInfoModel {
    activeDriver: boolean;
    drivingLicense: boolean;
    careerVision: string;
    presentation: string;
    areaOfInterests: number[];
    secondaryAreaOfInterests: number[];
    contractTypes: number[];
    preferedLocationId: string;
    willingToMove: boolean;
    preferredJobBeginning: Date;
}